+++ SVNavigatoru-Maven/src/main/java/com/svnavigatoru600/repository/users/impl/hibernate/UserDaoImpl.java
@@ -106,12 +106,19 @@ public class UserDaoImpl extends HibernateDaoSupport implements UserDao {
 
     @Override
     public void update(User user) {
+        this.update(user, true);
+    }
+
+    @Override
+    public void update(User user, boolean persistAuthorities) {
         this.logger.info("Update an user " + user);
         this.getHibernateTemplate().update(user);
 
-        // Updates authorities.
-        this.authorityDao.delete(user.getUsername());
-        this.authorityDao.save(user.getAuthorities());
+        if (persistAuthorities) {
+            // Updates authorities.
+            this.authorityDao.delete(user.getUsername());
+            this.authorityDao.save(user.getAuthorities());
+        }
     }
 
     @Override
